beta-version/requirements.txt
# Discord Music Bot - Production Dependencies
# Compatible with Python 3.10 - 3.13

# Discord
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0

# HTTP & Async
aiohttp>=3.9.0

# Audio Downloads
spotdl>=4.2.0
yt-dlp>=2024.01.01
gamdl>=1.8.0

# Lyrics
lyricsgenius>=3.0.1
syncedlyrics>=0.6.0

# Metadata & Media
mutagen>=1.47.0
Pillow>=10.0.0

# Configuration
python-dotenv>=1.0.0

# Caching
cachetools>=5.3.0

# Database
aiosqlite>=0.19.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# System Requirements:
# - FFmpeg 6.0+ (dengan libopus support)
# - Python 3.10 - 3.13
# - MP4Box (untuk gamdl)
# - mp4decrypt (untuk gamdl)

# Romanization for non-latin lyrics
pykakasi>=2.3.0      # Japanese romanization
pypinyin>=0.55.0     # Chinese pinyin
korean-romanizer>=0.28.0  # Korean romanization

# System monitoring
psutil>=5.9.0        # System and process utilities

# Translation & Language Support
googletrans==4.0.0rc1
deep-translator>=1.11.4

# Enhanced UI/UX
flask-minify>=0.42

# PWA Support (already have flask, just need manifest)

# Image processing (already have Pillow)

# Audio metadata enhancement (already have mutagen)
